refactor: 상품 관련 타입 및 카테고리 정의 파일 분리
- 상품 타입 및 카테고리 정의를 새로운 product.ts 파일로 이동하여 코드 구조 개선
- createProduct.ts에서 불필요한 타입 정의 제거로 가독성 향상
- Order 타입 및 OrderPageResponse 인터페이스 추가로 주문 관련 데이터 구조 정의
